Fixed the price box so you can now add decimal values for price

+++ src/main/java/com/makersacademy/nevermore/model/Cost.java
@@ -26,7 +26,7 @@ public class Cost {
     private Long id;
 
     private String content;
-    private int price;
+    private double price;
     private String category;
     public Date date;
 
@@ -36,14 +36,14 @@ public class Cost {
 
     public Cost() {}
 
-    public Cost(String content, Integer price, String category) {
+    public Cost(String content, Double price, String category) {
         this.content = content;
         this.price = price;
         this.category = category;
         
     }
 
-    public Cost(String content, int price, String category, Date date) {
+    public Cost(String content, Double price, String category, Date date) {
         this.content = content;
         this.price = price;
         this.category = category;
@@ -53,7 +53,7 @@ public class Cost {
     
     //Getters
     public String getcontent() { return this.content; }
-    public Integer getPrice()  { return this.price;}
+    public double getPrice()  { return this.price;}
     public String getCategory() { return this.category;}
     public Long getId() { return this.id; }
     public Date getDate() {return this.date;}
@@ -67,7 +67,7 @@ public class Cost {
         this.user = user;
     }
     public void setContent(String content) { this.content = content; }
-    public void setPrice(Integer price) { this.price = price;}
+    public void setPrice(Double price) { this.price = price;}
     public void setCategory(String category) { this.category = category;}
     public void setDate(Date date) { this.date = date;}
 }
